Homework Objective
The eruption times for Old Faithful are clustered into two different groups. One group is short eruptions, and the other group is long eruptions. Short eruptions last three minutes or less, while long eruptions last more than three minutes. As below, define the vector eruptions.times as:

eruptions.times <- faithful$eruptions

Split the vector eruptions.times into two vectors: a vector short that contains the times less than or equal to three minutes, and a vector long that contains the times greater than three minutes. Answer the following questions:

How many elements are in the vector short?

How many elements are in the vector long?

What is the mean erruption time of the short eruptions?

What is the mean erruption time of the long eruptions?

What is the standard deviation of the short eruption times?

What is the standard deviation of the long eruption times?