Homework Objective
The built in data set rivers has the length in miles of 141 major rivers in North America. You can build a data frame of this data set that is suitable for graphing as follows:

rr = data.frame(rivers)
Using the following command will provide the first 6 rows of data in the data frame, it should look something like this:

head(rr)
  rivers

1    735
2    320
3    325
4    392
5    524
6    450
With this entire data set, create a histogram with suitable bin widths, a box plot, and a normal probability plot. Then answer the following questions:

Are there any outliers in this data set? Are they high or low outliers?

Do these data appear to come from a normal distribution?

Create a presentation (MS Power Point or equivalent) that answers these questions, and includes the graphs and code that you used.